Fix large clipboard copy/paste (#10670)
Re-assemble cliprdr messages that were split into multiple PDUs

Implement reassembly of chunked virtual channel data, as described in
https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-rdpbcgr/a542bf19-1c86-4c80-ab3e-61449653abf6

Note: this is technically applicable to all virtual channels, not just
cliprdr, so this logic should eventually be moved to a more reusable
place. We haven't had an issue with the rdpdr virtual channel because
the smart card messages we implement are all small enough to fit in a
single chunk.