Default to `https` scheme for `--proxy` argument in `tctl auth sign`
Fixes https://github.com/gravitational/cloud/issues/1358.

Before this PR, if `--proxy` was set, it would be passed as it to the
kubeconfig file. Due to this, if the `--proxy` URL did not have a
scheme, it would default to `http,  leading to the issue reported in
https://github.com/gravitational/cloud/issues/1358.

With this PR, we now try to parse the `--proxy` URL and set its scheme
to `https` in case it's not set.