RFD 56: gRPC Backend Proposal (#10541)
Design for implementing a new gRPC client backend for Teleport.

The main driving motivation for this proposal is growth and requirements driven from Cloud-Hosted Teleport rather than demand from external customers. The server side implementation is outside of scope for this document, but the Why? in the RFD touches on what is driving this proposal

The initial implementation is intended to be done by the Cloud-Reliability team (with oversight, review and assistance from Core)