Fix meaning of `bot_name` in bot join tokens
The `BotName` / `bot_name` was previously used to refer to the bot
_username_ rather than the bot name as entered by users. This leads
to confusion as the username is an implemention detail not obviously
visible to users.

This changes the meaning of the `bot_name` parameter to instead
consistently accept a bot name, which is converted to a username on
the backend where needed.