Remove LDAP password_file from configuration
When we deprecated the password_file option for Teleport 9, we left
the configuration property in the config so that we could give v8
users who had recently upgraded a nice error message letting them
know that we deprecated this field.

For Teleport 10, everyone coming from v9 will have already removed
this property, so the deprecation warning is no longer necessary.