Edit Helm installation instructions (#11303)
Backports #11177

While addressing #10636, I noticed that the Details box that appears
for Cloud users when they visit some of our Helm guides is in a slightly
odd location, in the middle of a Prerequisites item. I have moved this to
the top of each guide instead.

I have also refactored the helm-install.mdx partial to split it into
multiple partials, which allows us to use the "Step n/d" format in our
Helm guides.

Also adds a Cloud warning to the Digital Ocean Helm guide.