Clean up remoteSites with no active tunnels
Not closing a remoteSite causes a continuous retry to init the cache
on the remote auth client. Not only does this cause log spam but is
a waste of resources.

We now stop tracking and close a remoteSite when its heartBeat fails
and no active tunnels are established for the site.

Fixes #11197