Move channel emitter to new eventstest package
In an attempt to reduce the huge API surface area of lib/events,
we create a new package that holds lib/events implmentations that
are only used for tests.

Also remove a duplicate implementation. A centralized package for
lib/events testing utilities should hopefully prevent a bunch of
these springing up in various test files.