Fix drone docker build tagging
This fixes an issue where the oss and enterprise docker image builds use
the default value for the VERSION variable set in teleport/Makefile to
tag the docker images.

Now docker images will be tagged with the git tag falling back to the
default value of VERSION when the tag is empty.