Fix invalid YAML in Machine ID configuration reference (#11583)
This fixes a smaller error showing invalid YAML in the Machine ID
reference configuration. While we do support a shorthand directory
path as just a string, the two cannot be combined.

This changes the documented recommendation to always use the more
verbose form where `path` and other directory-specific options are
fields on a `directory` object.