Clean up remoteSites with no active tunnels (#11435)
Not closing a remoteSite causes a continuous retry to init the cache
on the remote auth client. Not only does this cause log spam but is
a waste of resources.

We now stop tracking and close a remoteSite when its heartBeat fails
and no active tunnels are established for the site.

Fixes #11197

(cherry picked from commit 3d5928f07713ae8b2499e82992f958d7c2785327)

# Conflicts:
#	lib/reversetunnel/fake.go
#	lib/srv/db/access_test.go