Prevent cache reinitialization caused by missing watch event
Any Teleport v6 components with a cache that registers to watch
`types.KindClusterAuthPreference` that are running in a cluster
with any Teleport > v7 components results in:

Re-init the cache on error: watcher is closed: resource type
*types.AuthPreferenceV2 is not supported.

To prevent caches from being unhealthy, this updates the WatchEvent
stream to include the AuthPreference message. This allows said
event to be sent and received without closing the stream and making
the cache reload.