Hide navigation menu items based on scope
Closes #11383

Ensure that no visitor to the Teleport docs site sees content that is
irrelevant to their scope (e.g., Cloud, Open Source, or Enterprise) by
hiding scope-irrelevant content from the navigation menu.

In some cases, e.g., the introduction pages for the Cloud, Getting
Started, and Enterprise sections, the content is irrelevant to certain
scopes but a reader still may want to find out more.

One possibility would be to leave these section as-is for readers with
an unintend scope. However, these readers would see edition warnings in
the pages within a section unless they changed the value of the scope
picker.

As an alternative, this change used ScopedBlock to display links to
the relevant scope for users with a scope that is not intended for
a given menu page.

Also directs Enterprise visitors to the Getting Started menu page
to the Enterprise Getting Started page with an "enterprise" scope.
This fixes #10594.

Finally, wherever we include a warning at the top of an
edition-incompatible guide, I've added links to the scopes that
are supported for that guide. This should make it easier for
users to realize that they can adjust the scope to have the intended
docs experience.