Backport #10759 to branch/v9 (#11966)
* Edit three Setup guides for Cloud users

Teleport Daemon

- Use a Notice for a warning re: the Auth Service scoped to
  self-hosted users
- Misc grammar/style/clarity tweaks

Move the Enterprise License File page

The page does not belong in the Setup section, since it does not
apply to Cloud and OSS users. I have moved the page into the
Enterprise section.

Networking guide

- Add tabbed instructions so users who have selected a given scope
  don't see content for other scopes.

- Mention auth_service.proxy_listener_mode instead of
  tls_routing_enabled for self-hosted readers

- Misc grammar, style, and clarity edits

Scaling

  We already added an edition warning in a Details box, but I thought
  a Notice would be more prominent. I've also added some minor grammar,
  style, and clarity tweaks.

* Address PR feedback

- Imply that there are non-Auth/Proxy services beside Nodes
- Remove "shell" where it is unnecessary
- Revert the "HTTP CONNECT" section to where it was before this
  change. The copy-edits were inaccurate, but I don't have time
  to perform the edits agin more carefully.

* Address PR feedback

Simplify the HTTP CONNECT section by bringing the example closer
to the explanation of this functionality and removing some
ambiguity over whether one or both of HTTPS_PROXY and HTTP_PROXY are
required. The original text suggested that either is required, then
that both are required. The new text says to assign both, then
explains why.