Fix race on TestFIDO2Login/NOK_cancel_after_PIN (#12238)
Fixes a rather exoteric race condition where a goroutine beats an
already-cancelled context in a select.

The select in question sits on selectDevice / fido2.go#L729. It is not an error
for the select to go either way, instead this is a failure of test setup. The
 fix is simple: we use authenticators that do have resident credentials for the
 test, so their Assertion operations block waiting for user interaction,
 eliminating the race in the test in question.

Example flaky execution:
https://console.cloud.google.com/cloud-build/builds/ceaa0674-db49-444b-b059-a17e01b0328f?project=ci-account

#9160