Updated automatic backport creation workflow.
Removed automatic creation of backport Pull Requests due to limitations
within GitHub Actions on recursive creation of workflows not allowing
the "Check" workflow to run upon automation creation of the Pull
Request.

As a temporary workaround, prepare the backport branch and provide a
link the user can click on to create the backport Pull Request.