Add context.Context to GetReverseTunnels (#12373)
The RemoteClusterTunnelManager was hanging on the call to
GetReverseTunnels because it did not take a context and thus did
not respect cancellation. During certain shutdown/restart scenarios
the auth server may not respond, which causes Sync to never return
and subsequently prevents Run from ever completing.

By requiring a context to be passed to GetReverseTunnels, we can
remove the hard coded context.TODO that was being sent to the
http client and fully respect cancellation.

(cherry picked from commit 2c226600a75fd5657f62b4fa146bab812e25c187)

# Conflicts:
#	lib/auth/api.go
#	lib/auth/auth.go
#	lib/auth/auth_with_roles.go
#	lib/cache/cache.go