document and assert that detectProto needs at least 8 bytes
Both detectProto and isHTTP assume the input byte slice is only 3 bytes,
but both callers of detectProto always pass in 8 bytes. Also, some cases
truncate protocol prefixes that are larger than 3 bytes assuming that the
input is 3 bytes but other cases don't. To make things consistent and more
accurate, document that at least 8 bytes are required, and add a check for
good measure.