Backport #12179 to branch/v7 (#12368)
* Add a partial for agent installs in Teleport Cloud

If a user installs an agent version that is greater than the control
plane version, the agent will not be able to contact the control plane
(e.g., it will call a gRPC method that the control plane doesn't
recognize). The current partial for installing Teleport on a Linux host
includes only instructions for installing the latest version. Since
Teleport Cloud is behind the latest version, I have added a partial for
installing a Cloud-compatible agent and included it for Cloud users
in guides that require installing a Teleport agent.

* Backport the cloud version variable

Using "9.1.2" since the Cloud version is the same regardless of
self-hosted version.