Prevent ListResources from ignoring limit (#12501)
ListResources was missing a check on whether or not the limit was
reached after each possible allowed resource was added to the page.
This manifests when certain resources are filter out by deny rules,
causing the page size check at the end of IterateResources to never
be correct. In this scenario, the entire set of resources will be
returned. This can lead to resource exhaustion and sending messages
that exceed the gRPC limit. As a result of this the api client will
keep issuing more ListResources requests with a smaller limit until
the limit reaches 1.

To prevent the scenario above, ListResources now matches it's companion
filterAndListNodes. Whenever an item is added to the returned page,
we now check if the limit has been reached, instead of only after
processing an entire page retrieved from the cache.