Adds optional deployment key for CI (#10506) (#12590)
There are private copies of the teleport repo that require the use of
deployment keys to perform operations on the code repository (e.g.
`git fetch --unshallow`).

This patch adds the ability for a GCB build trigger to optionally pass
in the location of a secret deployment key, which will be fetched and
used if supplied.

If no such deployment key is supplied (i.e. in the public teleport repo),
the build will have access to neither the location of the deployment key,
nor the key itself, and thus cannot leak it during CI.

This patch also pulls the unshallow operation into the main build script (as
opposed to being a separate build step) so it can use the deployment key
to grant access to the remote repo.

Most of the new interaction with the repository is via shelling out to the
command-line git installed on the buildbox. My original intention was to
use go-git to manage the unshallow directly from code, but this was not
supported by the library.