Reduce latency of GetNodes (#12617)
Improves time to load nodes by retrieving the users role set
prior to iterating through a page of server resources. ListNodes
was pulling the role set once per page. List Resources was pulling
the role set once per server per page. Both of these resulted in
significant time to load nodes, which caused the web UI to be slower
to render the nodes page and to connect to a node.

Fixes #12599