Optionally skip unshallowing step
GCB checks out code for testing a PR as a shallow clone. The `teleport`  build requires unshallowing the clone in order to run analytics on the change, so it runs `git fetch --unshallow` to get the whole repo.

Unfortunately, the `teleport.e` repo already has a full teleport repository when it runs the test scripts, so unshallowing fails.

This change allows the unit & integration testing scripts to optionally skip the unshallow step, so that the teleport.e build triggers can succeed.