Add `session.start` and `session.end` events for non-interactive kube exec
This commit adds the `session.start` and `session.end` events to the non interactive exec commands.
Before this commit the `Exec` event was not recorded in Audit Log because it missed the `Metadata.Code` field and failed the validation.
This code is only known after the exec is triggered (FAILURE or SUCCESS) and so it requires that `Exec` event insertion into Audit Log  must be delayed until the execution finishes (i.e. like in SSH).
Given that, in order to prevent delayed log entries (execution is blocked for long time) we require the usage of  `SessionStart` and `SessionEnd` events even if the session is non-interactive.

Fixes #12764